# cmd/docs/swagger.yaml
definitions:
...
  PlantSite_internal_api_search-api_mapper.SearchPlantsItem:
    properties:
      params:
        additionalProperties: true
        type: object
      type:
        type: string
    required:
    - params
    - type
    type: object
  PlantSite_internal_api_search-api_mapper.SearchPostsItem:
    properties:
      params:
        additionalProperties: true
        type: object
      type:
        type: string
    required:
    - params
    - type
    type: object
  PlantSite_internal_api_search-api_response.GetPlantPhoto:
    properties:
      description:
        type: string
      id:
        type: string
      key:
        type: string
    required:
    - description
    - id
    - key
    type: object
  PlantSite_internal_api_search-api_response.GetPlantResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      latin_name:
        type: string
      main_photo_key:
        type: string
      name:
        type: string
      photos:
        items:
          $ref: '#/definitions/PlantSite_internal_api_search-api_response.GetPlantPhoto'
        type: array
      specification: {}
    required:
    - category
    - created_at
    - description
    - id
    - latin_name
    - main_photo_key
    - name
    type: object
  PlantSite_internal_api_search-api_response.GetPostPhoto:
    properties:
      id:
        type: string
      key:
        type: string
      place_number:
        type: integer
    type: object
  PlantSite_internal_api_search-api_response.GetPostResponse:
    properties:
      author_id:
        type: string
      content:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      id:
        type: string
      photos:
        items:
          $ref: '#/definitions/PlantSite_internal_api_search-api_response.GetPostPhoto'
        type: array
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  PlantSite_internal_api_search-api_response.SearchPlantItem:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      latin_name:
        type: string
      main_photo_key:
        type: string
      name:
        type: string
      specification: {}
    required:
    - category
    - created_at
    - description
    - id
    - latin_name
    - main_photo_key
    - name
    - specification
    type: object
  PlantSite_internal_api_search-api_response.SearchPostItem:
    properties:
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      photos:
        items:
          $ref: '#/definitions/PlantSite_internal_api_search-api_response.SearchPostPhoto'
        type: array
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    required:
    - author_id
    - content
    - created_at
    - id
    - title
    - updated_at
    type: object
  PlantSite_internal_api_search-api_response.SearchPostPhoto:
    properties:
      id:
        type: string
      key:
        type: string
      place_number:
        type: integer
    required:
    - id
    - key
    - place_number
    type: object
...
info:
  contact: {}
paths:
...
  /search/plant/{id}:
    get:
      description: Gets a plant by ID
      parameters:
      - description: Plant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Plant fetch successfully
          schema:
            $ref: '#/definitions/PlantSite_internal_api_search-api_response.GetPlantResponse'
        "400":
          description: Bad Request - Invalid input
        "401":
          description: Unauthorized - Not authorized to get plant
        "403":
          description: Forbidden - Does not have author rights to get plant
        "500":
          description: Internal Server Error - Failed to get plant
      summary: Get plant
      tags:
      - search
  /search/plants:
    post:
      consumes:
      - application/json
      description: Search plants using an array of different filter types
      parameters:
      - description: Array of search filters
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/PlantSite_internal_api_search-api_mapper.SearchPlantsItem'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/PlantSite_internal_api_search-api_response.SearchPlantItem'
            type: array
        "400":
          description: Invalid request format or missing required fields
        "500":
          description: Internal server error
      summary: Search plants with multiple filters
      tags:
      - search
  /search/post/{id}:
    get:
      description: Gets a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post fetch successfully
          schema:
            $ref: '#/definitions/PlantSite_internal_api_search-api_response.GetPostResponse'
        "400":
          description: Bad Request - Invalid input
        "401":
          description: Unauthorized - Not authorized to get post
        "403":
          description: Forbidden - Does not have author rights to get post
        "500":
          description: Internal Server Error - Failed to get post
      summary: Get post
      tags:
      - search
  /search/posts:
    post:
      consumes:
      - application/json
      description: Search posts using an array of different filter types
      parameters:
      - description: Array of search filters
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/PlantSite_internal_api_search-api_mapper.SearchPostsItem'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/PlantSite_internal_api_search-api_response.SearchPostItem'
            type: array
        "400":
          description: Invalid request format or missing required fields
        "500":
          description: Internal server error
      summary: Search posts with multiple filters
      tags:
      - search
swagger: "2.0"